Added highlighting when in interactivity range
When the player is within 1 meter of the center of an interactable object, it lights up with a yellow outline/highlight and a "E - Interact" tooltip appears above it.
Added the Outline package to handle outlining/hightlighting.
The Interactible script handles checking if the object is being interacted with, and calls the correct script to handle the interaction based on its type (an enum).